+++ concrete_datastore/api/v1_1/serializers.py
@@ -20,6 +20,12 @@ from concrete_datastore.concrete.meta import get_meta_definition_by_model_name
 concrete = apps.get_app_config('concrete')
 
 
+class ProcessRegisterSerializer(serializers.Serializer):
+    application = serializers.CharField(max_length=200, required=True)
+    instance = serializers.CharField(max_length=200, required=True)
+    token = serializers.UUIDField(required=True)
+
+
 def make_account_me_serialier(api_namespace=DEFAULT_API_NAMESPACE):
     meta = get_meta_definition_by_model_name('User')
     meta_serializer_class = make_serializer_class(